requirements.txt
fastapi
uvicorn
python-dotenv
python-binance
pandas
scikit-learn